Moved add product & cost cells into another module to reduce the line count of the other cells modules
