+++ src/components/CostAndPricing/PricingTab/cells/index.jsx
@@ -1,3 +1,4 @@
 export * from './footers';
 export * from './headers';
 export * from './regular';
+export * from './special';
