+++ src/components/CostAndPricing/PricingTab/cells/regular.jsx
@@ -1,23 +1,13 @@
 // @ts-check
 
-import { Add, Calculate, Cancel, Clear, Delete } from '@mui/icons-material';
-import {
-  Button,
-  ButtonGroup,
-  DialogContentText,
-  IconButton,
-  InputAdornment,
-  TextField,
-  Tooltip,
-} from '@mui/material';
+import { Calculate, Clear, Delete } from '@mui/icons-material';
+import { IconButton, InputAdornment, TextField, Tooltip } from '@mui/material';
 import { produce } from 'immer';
 import { useEffect, useState } from 'react';
-import { BaseDialog } from '../dialogs';
 import { ConfirmButtonDialog } from '../dialogs-wrapped';
 import * as fmt from '../formats';
 import { NumericInput } from '../inputs';
-import { PricingTableFab as Fab, TableTextField } from '../stylized';
-import { toCostNames, unique } from '../utils';
+import { TableTextField } from '../stylized';
 
 /**
  * A component that renders editable cells with dynamic costs
@@ -295,91 +285,4 @@ export function PercentCell({ getValue }) {
   const value = getValue();
   const percent = Number(value);
   return fmt.percent(percent);
-}
-
-/**
- * @param {import('../prop-types').AddProductCellProps} props
- */
-export function AddProductCell({ table }) {
-  const [open, setOpen] = useState(false);
-  const [productName, setProductName] = useState('');
-
-  const onClose = () => {
-    setOpen(false);
-    setProductName('');
-  };
-
-  const addProduct = () => {
-    table.options.meta?.setQuote(
-      produce((/** @type {import('../data-types').Quote} */ draft) => {
-        // This is probably the safest way to get a unique list of cost names.
-        const costNames = draft.products.flatMap(toCostNames).filter(unique);
-
-        // This *should* also work, but it might not order things correctly,
-        // and we need to update the tsconfig/jsconfig to iterate through the
-        // values.
-        // const costs = new Set(
-        //   draft.products.flatMap((p) => p.costs.map((c) => c.name)),
-        // );
-
-        draft.products.push({
-          name: productName,
-          quantity: 0,
-          selling_price_per_unit: 0,
-          estimated_hours: 0,
-          costs: costNames.map((name) => ({ name, value: 0 })),
-        });
-      }),
-    );
-  };
-
-  const onSubmit = () => {
-    addProduct();
-    onClose();
-  };
-
-  return (
-    <>
-      <Tooltip title="Add Product" arrow>
-        <Fab
-          size="small"
-          aria-label="Add Product"
-          onClick={() => setOpen(true)}
-        >
-          <Add />
-        </Fab>
-      </Tooltip>
-      <BaseDialog
-        open={open}
-        title="Product Name"
-        actions={
-          <ButtonGroup variant="contained">
-            <Button color="secondary" onClick={onClose} startIcon={<Cancel />}>
-              Cancel
-            </Button>
-            <Button type="submit" startIcon={<Add />}>
-              Add Product
-            </Button>
-          </ButtonGroup>
-        }
-        onClose={onClose}
-        onSubmit={onSubmit}
-      >
-        <DialogContentText>
-          Please specify the name of the new product.
-        </DialogContentText>
-        <TextField
-          autoFocus
-          required
-          margin="dense"
-          id="productName"
-          name="productName"
-          label="Product Name"
-          fullWidth
-          value={productName}
-          onChange={(e) => setProductName(e.target.value)}
-        />
-      </BaseDialog>
-    </>
-  );
 }
